LRCI-2795 Import into Testray using both S3 & RSYNC if available
