+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/DXPCloudClientTestrayImporter.java
@@ -334,7 +334,7 @@ public class DXPCloudClientTestrayImporter {
 		}
 
 		TestrayServer testrayServer = TestrayFactory.newTestrayServer(
-			_testrayServerURL, "S3");
+			_testrayServerURL);
 
 		TestrayProject testrayProject = testrayServer.getTestrayProjectByName(
 			_testrayProjectName);
