+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/TestrayFactory.java
@@ -128,32 +128,15 @@ public class TestrayFactory {
 	}
 
 	public static TestrayServer newTestrayServer(String testrayServerURL) {
-		return newTestrayServer(testrayServerURL, "RSYNC");
-	}
-
-	public static TestrayServer newTestrayServer(
-		String testrayServerURL, String testrayServerType) {
-
-		if (JenkinsResultsParserUtil.isNullOrEmpty(testrayServerType)) {
-			testrayServerType = "RSYNC";
-		}
-
-		String key = testrayServerURL + "_" + testrayServerType;
-
-		TestrayServer testrayServer = _testrayServers.get(key);
+		TestrayServer testrayServer = _testrayServers.get(testrayServerURL);
 
 		if (testrayServer != null) {
 			return testrayServer;
 		}
 
-		if (testrayServerType.equals("S3")) {
-			testrayServer = new S3TestrayServer(testrayServerURL);
-		}
-		else {
-			testrayServer = new RsyncTestrayServer(testrayServerURL);
-		}
+		testrayServer = new DefaultTestrayServer(testrayServerURL);
 
-		_testrayServers.put(key, testrayServer);
+		_testrayServers.put(testrayServerURL, testrayServer);
 
 		return testrayServer;
 	}
